pyproject.toml
[project]
name = "shaggy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "pyside6>=6.9.2",
    "PyGObject==3.50.*",
    "zmq>=0.0.0",
    "numpy>=2.2.6",
    "torch>=2.4,<2.5",
    "protobuf>=6.20",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "pydantic>=2.12.5",
    "matplotlib>=3.10.8",
    "ipython>=8.37.0",
    "click>=8.3.1",
]

[tool.setuptools]
packages = ["shaggy"]

[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"
